+++ src/main/java/com/github/games647/lagmonitor/tasks/BlockingIODetectorTask.java
@@ -1,15 +1,24 @@
 package com.github.games647.lagmonitor.tasks;
 
 import com.github.games647.lagmonitor.LagMonitor;
+import com.github.games647.lagmonitor.PluginUtil;
+import com.google.common.collect.Sets;
 
 import java.lang.Thread.State;
+import java.util.Arrays;
+import java.util.Map.Entry;
+import java.util.Set;
 import java.util.TimerTask;
 import java.util.logging.Level;
+
+import org.bukkit.plugin.Plugin;
 
 public class BlockingIODetectorTask extends TimerTask {
 
     private final Thread mainThread;
     private final LagMonitor plugin;
+
+    private final Set<String> violatedPlugins = Sets.newHashSet();
 
     public BlockingIODetectorTask(LagMonitor plugin, Thread mainThread) {
         this.plugin = plugin;
@@ -33,14 +42,12 @@ public class BlockingIODetectorTask extends TimerTask {
                 if (isElementEqual(topElement, "java.net.DualStackPlainSocketImpl", "connect0")
                         || isElementEqual(topElement, "java.net.SocketInputStream", "socketRead0")
                         || isElementEqual(topElement, "java.net.SocketOutputStream", "socketWrite0")) {
-                    plugin.getLogger().log(Level.WARNING
-                            , "Server is performing socket operations on the main thread", new Throwable());
+                    logWarning("Server is performing socket operations on the main thread. Proparly caused by {0}");
                 } //File (in) - java.io.FileInputStream.readBytes
                 //File (out) - java.io.FileOutputStream.writeBytes
                 else if (isElementEqual(topElement, "java.io.FileInputStream", "readBytes")
                         || isElementEqual(topElement, "java.io.FileOutputStream", "writeBytes")) {
-                    plugin.getLogger().log(Level.WARNING
-                            , "Server is performing file operations on the main thread", new Throwable());
+                    logWarning("Server is performing file operations on the main thread. Proparly caused by {0}");
                 }
             }
         }
@@ -48,5 +55,33 @@ public class BlockingIODetectorTask extends TimerTask {
 
     private boolean isElementEqual(StackTraceElement traceElement, String className, String methodName) {
         return traceElement.getClassName().equals(className) && traceElement.getMethodName().equals(methodName);
+    }
+
+    private void logWarning(String message) {
+        Exception stackTraceCreator = new Exception();
+        StackTraceElement[] stackTrace = stackTraceCreator.getStackTrace();
+
+        //remove the parts from LagMonitor
+        StackTraceElement[] copyOfRange = Arrays.copyOfRange(stackTrace, 2, stackTrace.length);
+        Entry<Plugin, StackTraceElement> foundPlugin = PluginUtil.findPlugin(copyOfRange);
+        String pluginName = "unknown";
+        if (foundPlugin != null) {
+            pluginName = foundPlugin.getKey().getName();
+            if (!violatedPlugins.add(pluginName) && plugin.getConfig().getBoolean("oncePerPlugin")) {
+                return;
+            }
+        }
+
+        plugin.getLogger().log(Level.WARNING, message, pluginName);
+
+        if (plugin.getConfig().getBoolean("hideStacktrace")) {
+            if (foundPlugin != null) {
+                StackTraceElement source = foundPlugin.getValue();
+                plugin.getLogger().log(Level.WARNING, "Source: {0}, method {1}, line{2}"
+                        , new Object[]{source.getClassName(), source.getMethodName(), source.getLineNumber()});
+            }
+        } else {
+            plugin.getLogger().log(Level.WARNING, "", stackTraceCreator);
+        }
     }
 }